% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citmre-package.R
\docType{package}
\name{citmre-package}
\alias{citmre}
\alias{citmre-package}
\title{citmre: Colombian Index Tool Market Rate Exchange}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Downloads the Representative Market Rate Exchange (RMRE) from the <www.datos.gov.co> source. Allows setting the data series in time frequencies, splitting the time series through start and end functions, transforming the data set in log returns or levels, and making a Dynamic graph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavRodEcon/citmre}
  \item \url{https://davrodecon.github.io/citmre/}
  \item Report bugs at \url{https://github.com/DavRodEcon/citmre/issues}
}

}
\author{
\strong{Maintainer}: David Rodríguez \email{davestss@hotmail.com} (\href{https://orcid.org/0000-0001-5430-0787}{ORCID})

Authors:
\itemize{
  \item Sergio Sierra-Luján \email{sergiochess95@gmail.com} (\href{https://orcid.org/0000-0003-4055-8810}{ORCID}) [copyright holder]
}

}
\keyword{internal}
