% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citmre-package.R
\docType{package}
\name{citmre-package}
\alias{citmre}
\alias{citmre-package}
\title{citmre: cITMre (Colombian Index Tool Market Rate Exchange)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Colombian Index Tool (Market Rate Exchange) package downloads the Colombian Market Rate from the source: Portal de Datos Abiertos <www.datos.gov.co>, for direct use in R.
}
\author{
\strong{Maintainer}: David Rodríguez \email{davestss@hotmail.com} (\href{https://orcid.org/0000-0001-5430-0787}{ORCID})

Authors:
\itemize{
  \item Sergio Sierra-Luján \email{sergiochess95@gmail.com} (\href{https://orcid.org/0000-0003-4055-8810}{ORCID}) [copyright holder]
}

}
\keyword{internal}
